requirements.txt
python-i18n[YAML]
dash
pandas
plotly
pickle